requirements.txt
pandas>=2.0.0
numpy>=1.24.0
torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
decord>=0.6.0
facenet-pytorch>=2.5.3
opencv-python-headless>=4.8.0
timm>=0.9.0
Pillow>=10.0.0
